requirements.txt
django
django-ckeditor
Pillow
djangorestframework
feedparser
numpy
argon2-cffi
pdfkit
pandas
matplotlib
seaborn
django-mathfilters
requests
1